+++ src/main/java/dev/woori/wooriLearn/domain/user/service/UserService.java
@@ -9,7 +9,6 @@ import dev.woori.wooriLearn.domain.account.repository.PointsHistoryRepository;
 import dev.woori.wooriLearn.domain.auth.entity.AuthUsers;
 import dev.woori.wooriLearn.domain.auth.port.AuthUserPort;
 import dev.woori.wooriLearn.domain.user.dto.ChangeNicknameReqDto;
-import dev.woori.wooriLearn.domain.auth.dto.ChangePasswdReqDto;
 import dev.woori.wooriLearn.domain.user.dto.SignupReqDto;
 import dev.woori.wooriLearn.domain.auth.entity.Role;
 import dev.woori.wooriLearn.domain.user.dto.UserInfoResDto;
@@ -58,8 +57,6 @@ public class UserService {
 
         authUserRepository.save(authUser);
         userRepository.save(user);
-
-        // TODO: 회원가입 이후 기본 포인트를 지급한다 하면 포인트 내역 추가 로직 필요
         pointsHistoryRepository.save(
                 PointsHistory.builder()
                         .user(user)
