+++ src/test/java/dev/woori/wooriLearn/domain/user/UserServiceTest.java
@@ -79,68 +79,4 @@ class UserServiceTest {
         assertEquals("newNick", user.getNickname());
         verify(userRepository).findByUserId("testUser");
     }
-
-
-    @Test
-    void testChangePassword() {
-        AuthUsers auth = AuthUsers.builder()
-                .userId("testUser")
-                .password("oldHashedPw")
-                .role(Role.ROLE_USER)
-                .build();
-
-        ChangePasswdReqDto req = new ChangePasswdReqDto("currentPw", "newPw");
-
-        when(authUserRepository.findByUserId("testUser"))
-                .thenReturn(Optional.of(auth));
-        when(passwordEncoder.matches("currentPw", "oldHashedPw"))
-                .thenReturn(true);
-        when(passwordEncoder.encode("newPw"))
-                .thenReturn("newHashedPw");
-
-        // when
-        authService.changePassword("testUser", req);
-
-        // then
-        verify(authUserRepository).findByUserId("testUser");
-        verify(passwordEncoder).matches("currentPw", "oldHashedPw");
-        verify(passwordEncoder).encode("newPw");
-        assertEquals("newHashedPw", auth.getPassword());
-    }
-
-    @Test
-    void testChangePassword_WrongCurrentPassword() {
-        AuthUsers auth = AuthUsers.builder()
-                .userId("testUser")
-                .password("oldHashedPw")
-                .role(Role.ROLE_USER)
-                .build();
-
-        ChangePasswdReqDto req = new ChangePasswdReqDto("wrongPw", "newPw");
-
-        when(authUserRepository.findByUserId("testUser"))
-                .thenReturn(Optional.of(auth));
-        when(passwordEncoder.matches("wrongPw", "oldHashedPw"))
-                .thenReturn(false);
-
-        // expect
-        assertThrows(CommonException.class,
-                () -> authService.changePassword("testUser", req));
-
-        verify(passwordEncoder, never()).encode(any());
-    }
-
-    @Test
-    void testChangePassword_UserNotFound() {
-        ChangePasswdReqDto req = new ChangePasswdReqDto("currentPw", "newPw");
-
-        when(authUserRepository.findByUserId("nonExistUser"))
-                .thenReturn(Optional.empty());
-
-        assertThrows(CommonException.class,
-                () -> authService.changePassword("nonExistUser", req));
-
-        verify(passwordEncoder, never()).matches(any(), any());
-        verify(passwordEncoder, never()).encode(any());
-    }
 }
